Simplify Inherited Proofing VA Form using deco_lite gem
- Update spec to reflect changes.

changelog: Improvements, Upcoming Features, Simplify Inherited Proofing VA Form using deco_lite gem
