+++ app/forms/idv/inherited_proofing/va/form.rb
@@ -1,20 +1,35 @@
 module Idv
   module InheritedProofing
     module Va
-      class Form < Idv::InheritedProofing::BaseForm
-        class << self
-          def required_fields
-            @required_fields ||= %i[first_name last_name birth_date ssn address_street address_zip]
-          end
+      class Form < DecoLite::Model
+        validates :first_name, :last_name, :birth_date, :ssn, :address_street, :address_zip,
+                  presence: true
 
-          def optional_fields
-            @optional_fields ||= %i[phone address_street2 address_city address_state
-                                    address_country]
-          end
+        attr_reader :payload_hash
+
+        def initialize(payload_hash:)
+          raise ArgumentError, 'payload_hash is blank?' if payload_hash.blank?
+          raise ArgumentError, 'payload_hash is not a Hash' unless payload_hash.is_a? Hash
+
+          @payload_hash = payload_hash.dup
+
+          super hash: @payload_hash
         end
 
-        validates(*required_fields, presence: true)
+        def submit
+          validate
 
+          FormResponse.new(
+            success: valid?,
+            errors: errors,
+            extra: {},
+          )
+        end
+
+        # Perhaps overkill, but a mapper service of some kind, not bound to this class,
+        # that takes into consideration context, may be more suitable. In the meantime,
+        # simply return a Hash suitable to place into flow_session[:pii_from_user] in
+        # our inherited proofing flow steps.
         def user_pii
           raise 'User PII is invalid' unless valid?
 
@@ -30,6 +45,10 @@ module Idv
           user_pii[:zipcode] = address_zip
           user_pii
         end
+
+        private
+
+        attr_writer :payload_hash
       end
     end
   end
